extracted fetchAPI functions to external file and import to App.js
